GUI: define (a platform-specific) `idle()`, and hide timer handling in `win32.cpp`
Before commit `f0104c029772e70e8a7cec2064f9f79cd90a34c0`,
`FstPlugin::gui_idle()` was responsible for periodically updating the
GUI. This was reverted back to the original, Win32 `SetTimer()` based
implementation, because VST 3 didn't seem to have a similar GUI idle
mechanism (not without VSTGUI), so it was easier to just use the
platform-defined timer mechanism in both plugin types.

But this approach seems to be platform-specific, so it's better moved
into `win32.cpp`.

Since event handling in X11 (Linux) works in a fundamentally different
way from Win32's approach, we need to make room for implementing it.
The biggest difference is that in Win32, an event handler function (the
window procedure) can be registered for each widget, which is then
called for each new event by the platform, while in X11, the events are
read from a stream by the main loop of the application, and the platform
doesn't care how they are dispatched to the event handling code. But
there can be only one main loop in a thread, so we need to share event
handling with the plugin host somehow.

VST 3 has the Linux-specific `IRunLoop` interface for this (and also,
for timers), but for VST 2 (FST), there doesn't seem to be any other way
to extend the host's main loop than to handle the `effEditIdle` message,
and read the X11 event queue there.

Due to this, `GUI::idle()` will also need different implementations for
different platforms.