vst3: fix compiler warning in `Vst3Plugin::Processor` (fixes #6)
If the `read_stream()` method ever would have been called with a
non-positive `max_size` value, it could have done unsafe things. But
this theoretical warning suggests reconsidering a more fundamental
problem: actually, there is no point in this method being a
general-purpose stream reader in the first place, receiving the buffer
size as a parameter, since it is only ever called with a constant value
(`Serializer::MAX_SIZE`). Premature abstraction is the other root of all
evil, so let's just rename it to `read_serialized_patch()`, and get rid
of the general-purpose `max_size` parameter.