Param: improve smoothness and responsivity of MIDI Controllers and Flexible Controllers
Regardless of any ongoing internal automations, we know the exact ratio
of `FloatParam` objects at the beginning of the block, so if we have new
controller data, we can cancel all previous target values, and start
converging to the new one with a duration which is based on the distance
of the current value and the target. This way convergence can be
sufficiently slow to feel smooth, but also, there will be no pile up of
rapidly changing controller events.