split up variable size rendering batches so that delay feedback never runs out of samples
When a host that uses variable size batches (e.g. FL Studio) wants to
render a longer batch after a shorter one, and the Echo effect's delay
time is set to an extermely short value (e.g. a few milliseconds,
less than the average batch size), then there might be an insufficient
number of samples in the feedback buffer from the previous rendering
round, so when the feedback buffer gets merged into the delay buffer,
then there might be sudden jumps in the signal, which then gets quickly
repeated by the effect due to the short delay time, producing even more
noise. To avoid the feedback buffer running out of samples, we need to
split up the bigger batch to chunks which don't exceed the size of the
previous batch.

To avoid having to duplicate this non-trivial logic, the common
rendering code was extracted from the two plugin types.