FstPlugin: comply with the host's expectation of the current program after a program change (closes #21)
Though actually switching to the new program only happens when the next
block of samples is rendered, if the host does not call `get_program()`
and co. after that, then it will consistently display the previous
program to the user after every program change. Thus, plugin methods
which are expected to operate on the current program (from the host's
point of view), need to use the `next_program` variable instead of
`bank.get_current_program_index()`.