Synth: initialize CC 64 to 0.0
Just like Channel Pressure (Aftertouch), the Sustain Pedal (when used as
a controller for a param) should only affect the sound of presets which
can utilize it if the pedal is actually present and engaged. The initial
sounds of the presets were designed with the pedal being released, and
that's how it should sound for a user who isn't using one.