FstPlugin: collect string based messages in a separate queue that is guaranteed to be written by a single thread
If both the audio and the GUI thread tries to push to the queue at the
same time (e.g. if `setParameter()` gets called in the audio thread
while the GUI is also trying to push to the queue) then a message might
get lost by getting overwritten, which is not a big deal for messages
that don't involve dynamic memory, but it should be avoided for messages
that do (ie. messages that use strings which might not do short string
optimizations).

So messages which might use dynamic memory should be kept in a seaparate
queue which is guaranteed to be never written by multiple threads
simultaneously.