Effects: implement a simple tape emulation
Some of the built-in presets already have a tape-like lo-fi, fluttering
echo, and some even allow huge pitch-dives via manipulating the delay
times of the effects, so it makes sense to also have a proper tape
machine emulation, with hiss, pitch dives, flutter and wow, etc.

The setup and the start-stop logic are the results of a lot of
experimentation, but now that they are in place, they could use some
unit testing and documenting.