Tape: remove the compressor, let `Distortion` handle the dynamic range
Even after experimenting with upward compression and expansion, it
doesn't seem to have the effect that I was looking for. It can be added
later if necessary - especially after I figure out how to put its
controls on the GUI, since it can be weird that when the tape effect is
turned on by a slight change of one of its knobs, it also causes a
sudden change in dynamics and gain. Unfortunately, I don't have the
capacity to do a major GUI redesign right now, and on the current one,
there's no space left for another compressor, so let's keep things
simple for now.