Synth: copy MPE member channel messages to the manager channel and update `README.md`
Though this violates the MPE specs, at the same time, it allows all
global parameters to be affected by the respective CC messages.

A standard-compliant way to achieve a similar effect would be to set the
MPE channel for the paraphonic voice parameters, but then polyphonic
expressions wouldn't be able to influence effects, global LFOs, etc.