presets: don't modify MPE settings when loading a preset
Since the MPE zone config is (intentionally) not cleared when changing
presets (similarly to tuning settings), switching to a preset which
contains MPE config would silently turn it on or modify the existing
settings, then loading a non-MPE preset would leave it turned on. This
could be confusing, especially if a user accidentally loads such a
preset, without being aware of the MPE related behaviour or having MPE
set up correctly on the input side.