Makefile: do not treat `deprecated-declarations` warnings as errors for VST 3 (issue #37)
The C++ 17 standard deprecated `std::wstring_convert` and friends, and
C++ 26 will remove them, but the VST 3 SDK still relies on the former.
Since the standard doesn't provide a replacement option, there's no
simple quick fix for the SDK that would be suitable for `lib/_patches`,
so the warning will be suppressed for now, and the upstream project was
notified.

References:

 * https://www.open-std.org/JTC1/SC22/WG21/docs/papers/2023/p2872r2.pdf
 * https://open-std.org/JTC1/SC22/WG21/docs/papers/2025/n5014.pdf (page 2368)
 * https://github.com/steinbergmedia/vst3_base/issues/9